+++ 2_download_applications.ps1
@@ -156,6 +156,7 @@ If($PSBoundParameters.ContainsKey('ApplicationsOverrideFile')){
 ## ================================
 . "$ScriptsPath\Symbols.ps1"
 . "$ScriptsPath\Environment.ps1"
+. "$ScriptsPath\SevenZipCmdlets.ps1"
 . "$ScriptsPath\SoftwareInventory.ps1"
 . "$ScriptsPath\WindowsUpdate.ps1"