+++ Scripts/Environment.ps1
@@ -2,7 +2,7 @@ Function Test-IsISE {
     <#
     .SYNOPSIS
     Determines if script running in ISE
-    
+
     .EXAMPLE
     Test-IsISE
     #>
@@ -20,7 +20,7 @@ Function Test-VSCode{
     <#
     .SYNOPSIS
     Determines if script running in VScode
-    
+
     .EXAMPLE
     Test-VSCode
     #>
@@ -68,7 +68,7 @@ function ConvertFrom-FixedColumnTable {
     <#
     .SYNOPSIS
         Converts string output to psobject
-    
+
     .DESCRIPTION
         Converts string output in table format (with header) to psobject
 
@@ -87,7 +87,7 @@ function ConvertFrom-FixedColumnTable {
     param(
       [Parameter(ValueFromPipeline)] $InputObject
     )
-    
+
     Begin {
         Set-StrictMode -Version 1
         $LineIndex = 0
@@ -100,17 +100,17 @@ function ConvertFrom-FixedColumnTable {
         foreach ($line in $lines) {
             ++$LineIndex
             Write-Verbose ("LINE [{1}]: {0}" -f $line,$LineIndex)
-            if ($LineIndex -eq 1) { 
+            if ($LineIndex -eq 1) {
                 # header line
-                $headerLine = $line 
+                $headerLine = $line
             }
-            elseif ($LineIndex -eq 2 ) { 
-                
+            elseif ($LineIndex -eq 2 ) {
+
                 # separator line
                 # Get the indices where the fields start.
                 $fieldStartIndex = [regex]::Matches($headerLine, '\b\S').Index
                 # Calculate the field lengths.
-                $fieldLengths = foreach ($i in 1..($fieldStartIndex.Count-1)) { 
+                $fieldLengths = foreach ($i in 1..($fieldStartIndex.Count-1)) {
                 $fieldStartIndex[$i] - $fieldStartIndex[$i - 1] - 1
                 }
                 # Get the column names
@@ -120,16 +120,16 @@ function ConvertFrom-FixedColumnTable {
                     } else {
                         $headerLine.Substring($fieldStartIndex[$i], $fieldLengths[$i]).Trim()
                     }
-                } 
+                }
             }
             else {
-               
+
                 $i = 0
                 # ordered helper hashtable for object constructions.
-                $ObjectHash = [ordered] @{} 
+                $ObjectHash = [ordered] @{}
                 foreach ($colName in $colNames) {
                     Write-Verbose ("COLUMN: {0}" -f $colName)
-                    $ObjectHash[$colName] = 
+                    $ObjectHash[$colName] =
                         if ($fieldStartIndex[$i] -lt $line.Length) {
                             if ($fieldLengths[$i] -and $fieldStartIndex[$i] + $fieldLengths[$i] -le $line.Length) {
                                 $line.Substring($fieldStartIndex[$i], $fieldLengths[$i]).Trim()
@@ -183,6 +183,66 @@ Function New-ItemPath {
 }
 
 
+Function Test-IsDuplicateDirectoryExists{
+    [CmdletBinding()]
+    param (
+        [Parameter(Mandatory = $true,Position=0)]
+        $DestinationPath,
+        [switch]$Passthru
+    )
+
+    $WorkingFolder = Split-Path $DestinationPath -Leaf
+    If(Get-ChildItem -Path $DestinationPath -Directory -Filter $WorkingFolder){
+        If($Passthru){
+            return $WorkingFolder
+        }Else{
+            return $true
+        }
+    }Else{
+        return $false
+    }
+}
+
+
+Function Move-ItemUpDirectory {
+
+    [CmdletBinding()]
+    param (
+        [Parameter(Mandatory = $true,Position=0)]
+        $CurrentPath
+    )
+
+    $UpOneDirectory = Split-Path (Get-Item -Path $CurrentPath).FullName -Parent
+    $CurrentRootFolder = Split-Path -Path $CurrentPath -Leaf
+    $Files = Get-ChildItem -Path $CurrentPath -Recurse -File
+    #TEST $File = $Files[0]
+    #TEST $File = $Files[2]
+    #TEST $File = $Files[-1]
+
+
+    foreach ($File in $Files) {
+
+        $CurrentFolders = (Split-Path -Path $File.FullName -Parent).replace("$UpOneDirectory\",'').replace("$CurrentRootFolder",'').TrimStart('\')
+        #$Destination_Path = Split-Path -Path $CurrentDirectory -Parent
+        $NewDirectory = Join-Path $UpOneDirectory -ChildPath $CurrentFolders
+        $OldDirectory = Join-Path $CurrentPath -ChildPath $CurrentFolders
+
+        If(-NOT(Test-Path -Path $NewDirectory -ErrorAction SilentlyContinue)){
+            New-Item -Path $NewDirectory -ItemType Directory -Force -ErrorAction SilentlyContinue | Out-Null
+        }
+
+        Move-Item -Path $File.FullName -Destination $NewDirectory -Force
+        if ($null -eq (Get-ChildItem -Path $OldDirectory -File -Recurse)) {
+
+            Remove-Item -Path $OldDirectory -Force | Out-Null
+        }
+    }
+
+    if ($null -eq (Get-ChildItem -Path $CurrentPath -File -Recurse)) {
+
+        Remove-Item -Path $CurrentPath -Force -Recurse | Out-Null
+    }
+}
 
 Function Get-GroupMembership {
     <#
@@ -235,7 +295,7 @@ Function Get-GroupMembership {
     }
 }
 
-function Get-LocalAdministrators {  
+function Get-LocalAdministrators {
     <#
     .SYNOPSIS
         Grabs local administrators
@@ -255,7 +315,7 @@ function Get-LocalAdministrators {
     [cmdletbinding()]
     Param(
         [string]$Computer=$env:computername
-    )  
+    )
 
     $CIMParams = @{
         Class='win32_groupuser'
@@ -267,19 +327,19 @@ function Get-LocalAdministrators {
     }
 
     Try{
-        $admins = Get-CimInstance @CIMParams  
-        $admins = $admins | Where-Object {$_.GroupComponent -like '*Administrators*'}  
+        $admins = Get-CimInstance @CIMParams
+        $admins = $admins | Where-Object {$_.GroupComponent -like '*Administrators*'}
 
-        $admins | ForEach-Object {  
-            $_.PartComponent -match ".+Name\ = (.+)\, Domain\ = (.+)$" > $null  
-            #$_.PartComponent -match ".+Domain\=(.+)\,Name\=(.+)$" > $null  
+        $admins | ForEach-Object {
+            $_.PartComponent -match ".+Name\ = (.+)\, Domain\ = (.+)$" > $null
+            #$_.PartComponent -match ".+Domain\=(.+)\,Name\=(.+)$" > $null
             $matches[2].trim(')').trim('"') + "\" + $matches[1].trim('"')
         }
     }
     Catch{
         Write-host ("Unable to Local Administrators. {0}" -f $_.Exception.Message) -ForegroundColor Red
         return $false
-    }  
+    }
 }
 
 Function Expand-StringVariables{
@@ -294,11 +354,11 @@ Function Expand-StringVariables{
         [switch]$ExpandProperty
 
     )
-    
+
     #convert settings object to hashtable to easily enumerate
     $ReplaceKeyValues = @{}
     $Object.psobject.properties | Where TypeNameOfValue -eq 'System.String' | Foreach { $ReplaceKeyValues[$_.Name] = $_.Value }
-    
+
     #Get all values within brackets
     $BracketVariables = ($Property |Select-String '(?<=\[)[^]]+(?=\])' -AllMatches).Matches.Value
     If($BracketVariables.count -gt 0){Write-Verbose ("Found bracket values: {0}" -f ($BracketVariables -join ','))}
@@ -308,9 +368,9 @@ Function Expand-StringVariables{
         #iterate through hashtable name and replace all matching keys with corrisponding values
         If($ReplaceKeyValues[$VariableName]){
             Write-Verbose ("Replacing bracket value [{0}] with object value: {1}" -f $VariableName,$ReplaceKeyValues[$VariableName])
-            $Property = $Property.replace("[$VariableName]", $ReplaceKeyValues[$VariableName]) 
+            $Property = $Property.replace("[$VariableName]", $ReplaceKeyValues[$VariableName])
         }
-            
+
         If($IncludeVariables){
             $VariableValue = (Get-Variable | Where Name -eq $VariableName).Value
             If($VariableValue){
@@ -334,7 +394,7 @@ Function Expand-StringVariables{
     }Else{
         return $Property
     }
-    
+
 }
 
 
@@ -349,7 +409,7 @@ Function ConvertTo-Variables{
     Process{
         foreach ($x in $object | get-member) {
             if ($x.MemberType -ne "Method" -and $x.Name -notlike "__*") {
-                
+
                 If(Get-Variable -Name $x.Name -ErrorAction SilentlyContinue){
                     Write-Verbose ("Updating variable name [{1}] to value [{2}] as type [{0}]" -f $x.Definition.Split(" ")[0],$x.Name,$object.$($x.Name))
                     Set-Variable -Name $x.Name -Value $object.$($x.Name) -Force
@@ -368,223 +428,6 @@ Function ConvertTo-Variables{
 
     }
 }
-
-
-
-Function Get-7zipUtilities{
-    Param(
-        [Parameter(Mandatory = $true)]
-        [string]$DestPath,
-        [switch]$Install
-    )
-
-    #Test if 7zip files exist
-    $sevenzipfiles = @(
-        "$ToolsPath\7za.exe" #standalone console version of 7-Zip with reduced formats support.
-        "$ToolsPath\7za.dll" #library for working with 7z archives
-        "$ToolsPath\7zxa.exe" #library for extracting from 7z archives
-        "$ToolsPath\7zxa.dll" #library for extracting from 7z archives
-    )
-
-    # Modern websites require TLS 1.2
-    [Net.ServicePointManager]::SecurityProtocol = [Net.SecurityProtocolType]::Tls12
-    
-    # Let's go directly to the website and see what it lists as the current version
-    $BaseUri = "https://www.7-zip.org/"
-    $BasePage = Invoke-WebRequest -Uri ( $BaseUri + 'download.html' ) -UseBasicParsing
-
-    # The most recent 'current' (non-beta/alpha) is listed at the top, so we only need the first.
-    If($Install){
-        $ChildPath = $BasePage.Links | Where-Object { $_.href -like '*7z*-x64.msi' } | Select-Object -First 1 | Select-Object -ExpandProperty href
-    }
-    Else{
-        $ChildPath = $BasePage.Links | Where-Object { $_.href -like '*7z*-extra.7z' } | Select-Object -First 1 | Select-Object -ExpandProperty href
-    }
-    
-    # Let's build the required download link
-    $DownloadUrl = $BaseUri + $ChildPath
-    
-    $OutFilePath = $DestPath + '\' + (Split-Path -Path $DownloadUrl -Leaf)
-
-    Write-Host "Downloading the latest 7-Zip to the temp folder"
-    Try{
-        Invoke-WebRequest -Uri $DownloadUrl -OutFile $OutFilePath -ErrorAction Stop | Out-Null
-    }
-    Catch{
-        Write-Host ("Unable to download 7-Zip. {0}" -f $_.Exception.Message) -ForegroundColor Red
-        return $false
-    }
-
-    Try{
-        If($Install){
-            Write-Host "Installing the latest 7-Zip"
-            Start-Process -FilePath "$env:SystemRoot\system32\msiexec.exe" -ArgumentList "/a", $OutFilePath -Wait 
-        }Else{
-            Add-Type -AssemblyName System.IO.Compression.FileSystem
-            [System.IO.Compression.ZipFile]::ExtractToDirectory($OutFilePath, $env:temp, $true)
-            #Expand-Archive -Path $OutFilePath -DestinationPath $DestPath -Force | Out-Null
-        }
-    }
-    Catch{
-        Write-Host ("Unable to setup 7-Zip. {0}" -f $_.Exception.Message) -ForegroundColor Red
-        return $false
-    }   
-}
-
-Function Compress-7zipArchive {
-    <#
-    .NOTES
-        -a Archive
-        -ai	Include the archive filenames
-        -an	Disable the parsing of the archive name
-        -ao	Overwrite mode
-        -ax	Exclude the archive filenames
-        -so	Write the data to StdOut
-        -si	Read the data from StdIn
-        -i	Include the filenames
-        -m	Set the compression method
-        -o	Set the output directory
-        -p	Set the password
-        -r	Recurse the subdirectories
-        -t	Type of archive
-        -u	Update the options
-        -v	Create the volumes
-        -w	Set the working directory
-        -x	Exclude the filenames
-        -y	Assume Yes on all the queries
-        -tzip
-        -t7z
-    #>
-    [CmdletBinding()]
-    param (
-        [Parameter(Mandatory = $true)]
-        $SevenZipPath,
-        
-        [Parameter(Mandatory = $true,ValueFromPipeline = $true)]
-        [Alias('Folder')]
-        [string[]] $Path,
-        
-        [Parameter(Mandatory = $true)]
-        [Alias('Destination')]
-        [string] $DestinationPath,
-        
-        [ValidateSet('100m','500m','1g')]
-        [String]$SplitSize,
-        
-        [switch] $Force,
-
-        [switch] $ShowProgress
-        
-    )
-    Begin{
-        $env:SEE_MASK_NOZONECHECKS = 1
-    }
-    Process{
-        $ZipArgs = @(
-            'a'
-            '-r'
-        )
-        If($SplitSize){$ZipArgs += "-v$SplitSize"}
-        If($Force){$ZipArgs += '-aoa'}
-        
-        if (-not (Test-Path -Path $SevenZipPath -PathType Leaf)) {
-            throw "7 zip file '$SevenZipPath' not found"
-        }
-
-        $ZipArgs = $ZipArgs -join ' '
-        Write-Verbose "Start-Process $SevenZipPath -ArgumentList `"$ZipArgs `"$DestinationPath`" `"$Path`" -PassThru -Wait -WindowStyle Hidden"
-        If($ShowProgress){
-            $result = Start-Process $SevenZipPath -ArgumentList "$ZipArgs `"$DestinationPath`" `"$Path`"" -RedirectStandardOutput "$env:temp\stdout.txt" -RedirectStandardError "$env:temp\stderr.txt"  -PassThru -WindowStyle Hidden
-            <#
-            $ErroredFiles = 0
-            #region Progress bar loop
-            while (!$AzCopy.HasExited) {
-                Start-Sleep $ReportGap
-                If($ShowProgress){
-                    $TransferStatus = Get-Content -Path "$env:temp\stdout.txt" | Select -Last 1
-                    If($TransferStatus -match '^\d+'){
-                        $DataSet = ($TransferStatus.split(',') -Replace '\w+$|%','')[0..4].Trim()
-                        If([int]$DataSet[2] -ne 0){$ErroredFiles=$DataSet[2]}
-                        Write-Progress -Activity ('Transferring files to [{0}]' -f $Destination) -Status ("Copied {0} of {1} files..." -f $DataSet[1], $DataSet[4]) -PercentComplete $DataSet[0]
-                    }ElseIf([string]::IsNullOrWhiteSpace($TransferStatus) ){
-                        Write-Progress -Activity ('AzCopy status' -f $Destination) -Status "Nothing to report" -PercentComplete 100
-                    }
-                    Else{
-                        Write-Progress -Activity ('AzCopy status' -f $Destination) -Status "$TransferStatus" -PercentComplete 100
-                    }
-                }
-            }
-            #>
-        
-        }Else{
-            $result = Start-Process $SevenZipPath -ArgumentList "$ZipArgs `"$DestinationPath`" `"$Path`"" -PassThru -Wait -WindowStyle Hidden
-        }
-        
-        
-        if ($result.ExitCode -eq 0) {
-            Write-Verbose "Folder archived successfully."
-        } else {
-            Write-Error ("Error occurred while archiving the folder. {0}" -f $result.ExitCode)
-        }
-    }
-    End{
-        $env:SEE_MASK_NOZONECHECKS = 0
-    }
-}
-
-Function Expand-7zipArchive{
-
-    [CmdletBinding()]
-    param (
-        [Parameter(Mandatory = $true)]
-        $SevenZipPath,
-        
-        [Parameter(Mandatory = $true,ValueFromPipeline = $true)]
-        [Alias('File')]
-        [string[]] $FilePath,
-        
-        [Alias('Destination')]
-        [string] $DestinationPath,
-
-        [switch] $Force,
-
-        [switch] $ShowProgress
-    )
-    Begin{
-        $env:SEE_MASK_NOZONECHECKS = 1
-    }
-    Process{
-        $ZipArgs = @('e')
-        If($Force){$ZipArgs += '-aoa'}
-        
-        #add first file path
-        $ZipArgs += "`"$FilePath`""
-    
-        #always make this last
-        $ZipArgs += '-o'
-        $ZipArgs = $ZipArgs -join ' '
-        
-        If($DestinationPath -notmatch '\*$'){$DestinationPath=$DestinationPath + '*'}
-
-        Write-Verbose "Start-Process $SevenZipPath -ArgumentList `"$ZipArgs`"$DestinationPath`"`" -PassThru -Wait -WindowStyle Hidden"
-        If($ShowProgress){
-            $result = Start-Process $SevenZipPath -ArgumentList "$ZipArgs`"$DestinationPath`"" -RedirectStandardOutput "$env:temp\stdout.txt" -RedirectStandardError "$env:temp\stderr.txt" -PassThru -WindowStyle Hidden
-        }Else{
-            $result = Start-Process $SevenZipPath -ArgumentList "$ZipArgs`"$DestinationPath`"" -PassThru -Wait -WindowStyle Hidden
-        }
-        
-        
-        if ($result.ExitCode -eq 0) {
-            Write-Verbose "Folder expanded successfully."
-        } else {
-            Write-Error ("Error occurred while expanding the folder. {0}" -f $result.ExitCode)
-        }
-    }
-    End{
-        $env:SEE_MASK_NOZONECHECKS = 0
-    }
-}
-
 
 
 # function to generate password; compatible with .Net core
@@ -625,7 +468,7 @@ function Test-IsPendingReboot {
     if (Get-ChildItem "HKLM:\Software\Microsoft\Windows\CurrentVersion\Component Based Servicing\RebootPending" -EA Ignore) { return $true }
     if (Get-Item "HKLM:\SOFTWARE\Microsoft\Windows\CurrentVersion\WindowsUpdate\Auto Update\RebootRequired" -EA Ignore) { return $true }
     if (Get-ItemProperty "HKLM:\SYSTEM\CurrentControlSet\Control\Session Manager" -Name PendingFileRenameOperations -EA Ignore) { return $true }
-    try { 
+    try {
         $util = [wmiclass]"\\.\root\ccm\clientsdk:CCM_ClientUtilities"
         $status = $util.DetermineIfRebootPending()
         if (($status -ne $null) -and $status.RebootPending) {