+++ Scripts/VM_Scripts/vm_install_applications.ps1
@@ -76,9 +76,9 @@ $FilteredCustomizations = ($ControlCustomizationData.customSequence | Where -Fil
 ## ================================
 . "$ScriptsPath\Symbols.ps1"
 . "$ScriptsPath\Environment.ps1"
+. "$ScriptsPath\SevenZipCmdlets.ps1"
 . "$ScriptsPath\SoftwareInventory.ps1"
 . "$ScriptsPath\WindowsUpdate.ps1"
-
 
 ## ================================
 ## IMPORT OFFLINE MODULES
@@ -89,12 +89,10 @@ Install-PackageProvider NuGet -Force
 #Register-PSRepository -Name Local -SourceLocation "$ToolsPath\Modules" -InstallationPolicy Trusted
 
 $OfflineModules = Get-ChildItem $ToolsPath -Recurse -Filter *.nupkg
-$ModulesNeeded = @('MSFTLinkDownloader','YetAnotherCMLogger','PSWindowsUpdate')
-#TEST $Module = $ToolkitSettings.Settings.offlineSupportingModules[0]
-#TEST $Module = $ToolkitSettings.Settings.offlineSupportingModules[2]
+$ModulesNeeded = @('PSWindowsUpdate')
 $i=0
-Foreach($Module in $ToolkitSettings.Settings.offlineSupportingModules){
-    Write-Host ("`n[{0} of {1}] Processing module [{2}]..." -f $i,$ToolkitSettings.Settings.offlineSupportingModules.count,$Module )
+Foreach($Module in $ModulesNeeded){
+    Write-Host ("`n[{0} of {1}] Processing module [{2}]..." -f $i,$ModulesNeeded.count,$Module )
     If($OfflineModule = $OfflineModules | Where Name -like "$Module*"){
 
         $Name = $OfflineModule.BaseName.split('.')[0].Trim()
@@ -121,16 +119,8 @@ Foreach($Module in $ToolkitSettings.Settings.offlineSupportingModules){
 ## MAIN
 ## ================================
 Write-Host ("`nSTARTING APP INSTALL PROCESS") -ForegroundColor Cyan
-$stopwatch = [System.Diagnostics.Stopwatch]::StartNew()
-
-#convert properties to variables (used durign expand tree)
-$ToolkitSettings | Select-Object -ExpandProperty TenantEnvironment | ConvertTo-Variables
 
 $i = 0
-#TEST $SequenceItem = $ControlCustomizationData.customSequence[9]
-#TEST $SequenceItem = $ControlCustomizationData.customSequence[0]
-#TEST $SequenceItem = $ControlCustomizationData.customSequence[13]
-#TEST $SequenceItem = $FilteredCustomizations[1]
 Foreach($SequenceItem in $FilteredCustomizations){
     $i++
     Write-Host ("`n[{0}/{1}] Processing step: {2} {3}..." -f $i,$FilteredCustomizations.count,$SequenceItem.name,$SequenceItem.version -replace '\[version\]','' )
@@ -235,7 +225,7 @@ Foreach($SequenceItem in $FilteredCustomizations){
                             If($Result.ExitCode -in $SequenceItem.ValidExitCodes){
                                 Write-Host "Install command ran successfully" -ForegroundColor Green
                             }Else{
-                                Write-YaCMLogEntry -Message ("Install command failed for [{0}], error [{1}]" -f $SequenceItem.name,$Result.ExitCode) -Severity 3
+                                #Write-YaCMLogEntry -Message ("Install command failed for [{0}], error [{1}]" -f $SequenceItem.name,$Result.ExitCode) -Severity 3
                                 Write-Host ('Failed. Exit Code: {0}' -f $Result.ExitCode) -ForegroundColor Red
                                 If([System.Convert]::ToBoolean($SequenceItem.continueOnError) -eq $false){
                                     Break
@@ -277,9 +267,6 @@ Foreach($SequenceItem in $FilteredCustomizations){
                         Break
                     }
 
-                    $stopwatch.Stop()
-                    $stopwatch.Reset()
-                    $stopwatch.Restart()
                 }#end filename loop
 
             }Elseif([System.Convert]::ToBoolean($SequenceItem.continueOnError)){
@@ -302,13 +289,14 @@ Foreach($SequenceItem in $FilteredCustomizations){
                 Try{
                     Invoke-Expression $expandedscript
                 }Catch{
-                    Write-YaCMLogEntry -Message ("Failed to run command [{0}], error [{1}]" -f $expandedscript,$_.exception.message) -Severity 3
+                    #Write-YaCMLogEntry -Message ("Failed to run command [{0}], error [{1}]" -f $expandedscript,$_.exception.message) -Severity 3
                     Write-Host ('Failed. Error: {0}' -f $_.exception.message) -ForegroundColor Red
                     If([System.Convert]::ToBoolean($SequenceItem.continueOnError) -eq $false){
                         Break
                     }
                 }
             }
+
             Write-Host ("Done") -ForegroundColor Green
         }#end script switch
 
@@ -338,6 +326,7 @@ Foreach($SequenceItem in $FilteredCustomizations){
                 }
                 Write-Host ("Done") -ForegroundColor Green
             }
+
         }#end windows update switch
     }